+++ beep/run_model.py
@@ -76,187 +76,6 @@ DEFAULT_MODEL_PROJECTS = [
 ]
 assert all("_" not in name for name in DEFAULT_MODEL_PROJECTS)
 
-FEATURIZER_CLASSES = [RPTdQdVFeatures, HPPCResistanceVoltageFeatures,
-                      HPPCRelaxationFeatures, DiagnosticSummaryStats]
-
-class Dataset(MSONable, metaclass=ABCMeta):
-    """
-    Class corresponding to a training dataset assembled from BeepFeatures objects
-
-    Attributes:
-
-    """
-
-    def __init__(self,name, data, metadata, filenames, feature_sets):
-        """
-
-        :param data: a combined dataframe
-        :param metadata: list of metadata dicts for the different feature objects
-        :param filenames: list of filenames that have atleast one of the feature objects
-        :param feature_sets: list of feature sets that were merged. This could be used
-        to group sets of features for techniques like grouped/hierarchical lasso etc
-        """
-        self.name = name
-        self.data = data
-        self.metadata = metadata
-        self.filenames = filenames
-        self.feature_sets = feature_sets
-
-    def as_dict(self):
-        """
-        Method for dictionary serialization
-        Returns:
-            dict: corresponding to dictionary for serialization
-        """
-        obj = {
-            "@module": self.__class__.__module__,
-            "@class": self.__class__.__name__,
-            "name": self.name,
-            "data": self.data.to_dict("list"),
-            "metadata": self.metadata,
-            "filenames": self.filenames,
-            "feature_sets": self.feature_sets
-        }
-        return obj
-
-    @classmethod
-    def from_dict(cls, d):
-        """MSONable deserialization method"""
-        d["data"] = pd.DataFrame(d["data"])
-        return cls(**d)
-
-
-    @classmethod
-    def from_features(cls, project_list, feature_class_list=FEATURIZER_CLASSES,
-                      feature_dir="data-share/features/")
-        """
-        Method to assemble a dataset from a list of BeepFeatures objects generated for one or more projects. 
-        Expected folder structure:
-
-        :param project_list:
-        :param feature_class_list:
-        :param feature_dir:
-        :param dataset_dir:
-        :return:
-        """
-        feature_df_list = []
-        metadata = []
-        feature_sets = {}
-        for feature_class in feature_class_list:
-            feature_df = pd.DataFrame()
-            for project in project_list:
-                feature_path = os.path.join(feature_dir, feature_class.class_feature_name)
-                feature_jsons = [f for f in os.listdir(feature_path) if
-                                 (os.path.isfile(os.path.join(feature_path, f)) and
-                                  f.startswith(project))]
-                for feature_json in feature_jsons:
-                    obj = loadfn(feature_json)
-                    df = obj.X
-                    df['file'] = obj.metadata['protocol']
-                    feature_df = pd.concat([feature_df, df]).reset_index(drop=True)
-                    ## TODO: Need some logic for ensuring that features of a given class being concatenated
-                    ## all have the same metadata dict
-
-            feature_df_list.append(feature_df)
-            feature_sets[feature_class] = list(feature_df.columns)
-
-        df = reduce(lambda x, y: pd.merge(x, y, on = 'file', how= 'inner'), feature_df_list)
-        return cls(df, metadata, df.file.unique(), feature_sets)
-
-    @classmethod
-    def from_processed_cycler_runs(cls, project_list, processed_run_list=None,
-                                   feature_class_list=FEATURIZER_CLASSES,
-                                  metadata_list=None, processed_dir="data-share/structure/")
-        """
-        Method to assemble a dataset from a list of ProcessedCyclerRun objects belonging to one or more projects.
-        Expected folder structure:
-
-        :param project_list: list of projects to featurize and combine as a training dataset
-        :param processed_run_list: list of paths to specific ProcessedCyclerRun objects to be featurized. 
-        If provided, this will over-ride project based looping.
-        :param feature_class_list: list of featurizers to invoke on the structured cycler files.
-        :param feature_dir: location to store serialized feature jsons
-        :param dataset_dir: location to store dataset
-        :return:
-        """
-        feature_df_list = [pd.DataFrame()]*len(feature_class_list)
-        feature_sets = {}
-
-        #Check if metadata is present, and if yes, check if the dictionary keys are right
-        if metadata_list is None:
-            print('No metadata specified for feature generation. Assuming defaults and proceeding.')
-            for idx, feature_class in enumerate(feature_class_list):
-                if feature_class.class_feature_name in FEATURE_HYPERPARAMS.keys():
-                    metadata_list[idx] = FEATURE_HYPERPARAMS[feature_class.class_feature_name]
-                else:
-                    metadata_list[idx] = None
-        else:
-            for idx, feature_class  in enumerate(feature_class_list):
-                if metadata_list[idx] is None:
-                    if feature_class.class_feature_name in FEATURE_HYPERPARAMS.keys():
-                        metadata_list[idx] = FEATURE_HYPERPARAMS[feature_class.class_feature_name]
-                    else:
-                        metadata_list[idx] = None
-                elif set(FEATURE_HYPERPARAMS[feature_class.class_feature_name.keys()]) != \
-                        set(metadata_list[idx].keys()):
-                    raise ValueError('Invalid hyperparameter dictionary')
-
-        if processed_run_list is None:
-            processed_run_list = [f
-                                  for f in os.listdir(processed_dir)
-                                  for project in project_list
-                                  if (os.path.isfile(os.path.join(processed_dir, f)) and
-                                      f.startswith(project))]
-
-        for processed_json in processed_run_list:
-            path = os.path.join(processed_dir, processed_json)
-            processed_cycler_run = loadfn(path)
-            for idx, feature_class in enumerate(feature_class_list):
-                obj = feature_class.from_run(
-                    path, '/data-share/features/', processed_cycler_run, metadata_list[idx])
-                df = obj.X
-                df['file'] = obj.protocol
-                feature_df_list[idx] = pd.concat([feature_df_list[idx], df])
-
-            for idx, feature_class in enumerate(feature_class_list):
-                feature_sets[feature_class] = list(feature_df_list[idx].columns)
-
-        df = reduce(lambda x, y: pd.merge(x, y, on = 'file', how= 'inner'), feature_df_list)
-        return cls(df, metadata_list, df.file.unique(), feature_sets)
-
-    def generate_train_test_split(self, predictors=None, outcomes=None,
-                                  split_by_cell=True, test_size=0.4, seed=123):
-        """
-
-        :param predictors: list of columns to use as predictors
-        :param outcomes: list of columns to use as outcomes
-        :param split_by_cell:
-        :param seed:
-        :return: X_train, X_test, y_train, y_test
-        """
-
-        if predictors is None:
-            raise ValueError('Specify one or more predictor columns')
-
-        if outcomes is None:
-            raise ValueError('Specify one or more outcomes')
-
-        np.random.seed(seed)
-        if split_by_cell:
-            test_cells = np.random.choice(self.filenames, int(len(self.filenames)*test_size))
-            train_cells = [x for x in self.filenames if x not in test_cells]
-
-            X_train = self.data.loc[self.data.file.isin(train_cells), predictors]
-            X_test = self.data.loc[self.data.file.isin(test_cells), predictors]
-
-            y_train = self.data.loc[self.data.file.isin(train_cells), outcomes]
-            y_test = self.data.loc[self.data.file.isin(test_cells), outcomes]
-
-            return X_train, X_test, y_train, y_test
-        else:
-            return train_test_split(self.data[predictors], self.data[outcomes], test_size, random_state=seed)
-
-
 class DegradationModel(MSONable):
     """
     Object corresponding to prediction model.
